Feat: Add support for OpenRouter
In order to enable more than just OpenAI as a backend, this commit adds
a new api_provider configuration option. This allows us to set openrouter
or openai as the provider and correctly set up our model through Raix.
Its worth mentioning that the default behavior of OpenAI is preserved
and this configuration option should be only additive.